requirements.txt
psycopg2-binary==2.9.10
selenium==4.30.0
python-dotenv==1.0.1
dotenv==0.9.9
requests>=2.28.0
google-cloud-storage>=2.0.0
google-api-python-client==2.136.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pandas==2.3.0